+++ src/ingest_validation_tools/__init__.py
@@ -1,5 +1,5 @@
 """Top-level package for ingest_validation_tools."""
 
 __author__ = """C McCallum"""
-__email__ = 'mccallucc@gmail.com'
-__version__ = '0.0.1'
+__email__ = "mccallucc@gmail.com"
+__version__ = "0.0.1"