+++ src/ingest_validation_tools/upload.py
@@ -17,16 +17,18 @@ from ingest_validation_tools.plugin_validator import run_plugin_validators_iter
 from ingest_validation_tools.schema_loader import (
     PreflightError,
     SchemaVersion,
+    get_table_schema,
 )
-from ingest_validation_tools.table_validator import ReportType
+from ingest_validation_tools.table_validator import ReportType, get_table_errors
 from ingest_validation_tools.validation_utils import (
+    TSVError,
     dict_reader_wrapper,
     get_context_of_decode_error,
     get_data_dir_errors,
     get_directory_schema_versions,
     get_json,
-    get_table_schema_version,
-    get_tsv_errors,
+    get_schema_version,
+    read_rows,
 )
 
 TSV_SUFFIX = "metadata.tsv"
@@ -76,7 +78,7 @@ class Upload:
 
         try:
             unsorted_effective_tsv_paths = {
-                str(path): get_table_schema_version(
+                str(path): get_schema_version(
                     path,
                     self.encoding,
                     self.globus_token,
@@ -93,7 +95,7 @@ class Upload:
             }
 
         except PreflightError as e:
-            self.errors["Preflight"] = str(e)
+            self.errors["Preflight"] = e
 
     # def _check_multi_assay(self):
     #     # This is not recursive, so if there are nested requirements it will not work
@@ -145,9 +147,10 @@ class Upload:
         ).strip()
 
         try:
+            # TODO: is there a particular way this needs to display?
             effective_tsvs = {
                 Path(path).name: {
-                    "Schema": sv.schema_name,
+                    "Schema": sv.table_schema,
                     "Metadata schema version": sv.version,
                     "Directory schema versions": get_directory_schema_versions(
                         path,
@@ -164,7 +167,7 @@ class Upload:
                 "TSVs": effective_tsvs,
             }
         except PreflightError as e:
-            self.errors["Preflight"] = str(e)
+            self.errors["Preflight"] = e
             return self.errors
 
     def get_errors(self, **kwargs) -> dict:
@@ -181,13 +184,13 @@ class Upload:
 
         errors = {}
         try:
-            other_errors = self.check_other_schemas()
-            if other_errors and not self.effective_tsv_paths:
-                return other_errors
-            elif self.effective_tsv_paths:
-                errors.update(other_errors)
-            else:
-                return {}
+            # other_errors = self.check_other_schemas()
+            # if other_errors and not self.effective_tsv_paths:
+            #     return other_errors
+            # elif self.effective_tsv_paths:
+            #     errors.update(other_errors)
+            # else:
+            #     return {}
 
             upload_errors = self.check_upload()
             if upload_errors:
@@ -220,21 +223,21 @@ class Upload:
     #
     ###################################
 
-    def check_other_schemas(self):
-        # Antibodies and contributors are handled elsewhere
-        errors = {}
-        others = ["organ", "sample"]
-        for tsv_path, schema_version in self.effective_tsv_paths.items():
-            if schema_version.schema_name in others:
-                error = self._validate(tsv_path, schema_version)
-                if error:
-                    errors[f"{tsv_path} (as {schema_version.schema_name})"] = error
-        self.effective_tsv_paths = {
-            k: v
-            for k, v in self.effective_tsv_paths.items()
-            if v.schema_name not in others
-        }
-        return errors
+    # def check_other_schemas(self):
+    #     # Antibodies and contributors are handled elsewhere
+    #     errors = {}
+    #     others = ["organ", "sample"]
+    #     for tsv_path, schema_version in self.effective_tsv_paths.items():
+    #         if schema_version.schema_name in others:
+    #             error = self._validate(tsv_path, schema_version)
+    #             if error:
+    #                 errors[f"{tsv_path} (as {schema_version.schema_name})"] = error
+    #     self.effective_tsv_paths = {
+    #         k: v
+    #         for k, v in self.effective_tsv_paths.items()
+    #         if v.schema_name not in others
+    #     }
+    #     return errors
 
     def check_upload(self) -> dict:
         upload_errors = {}
@@ -260,72 +263,99 @@ class Upload:
                     "Repeated": f"There is more than one TSV for this type: {', '.join(repeated)}"
                 }
             )
-        for path, schema_version in self.effective_tsv_paths.items():
-            rows = self._get_rows_from_tsv(path)
-            if not isinstance(rows, list):
-                return {f"{path} (as {schema_version.schema_name})": rows}
+        for path, schema in self.effective_tsv_paths.items():
+            if (
+                "data_path" not in schema.raw_rows[0]
+                or "contributors_path" not in schema.raw_rows[0]
+            ):
+                errors.update(
+                    {
+                        f"{path} (as {schema.table_schema})": [
+                            "File is missing data_path or contributors_path."
+                        ]
+                    }
+                )
             for ref in ["contributors", "antibodies"]:
-                ref_errors = self._get_ref_errors(rows, ref, schema_version, path)
+                ref_errors = self._get_ref_errors(ref, schema, path)
                 if ref_errors:
                     errors.update(ref_errors)
         return errors
 
     def _get_directory_errors(self) -> dict:
         errors = {}
-        for path, schema_version in self.effective_tsv_paths.items():
-            rows = self._get_rows_from_tsv(path)
-            if not isinstance(rows, list):
-                return {f"{path} (as {schema_version.schema_name})": rows}
-            if "data_path" not in rows[0] or "contributors_path" not in rows[0]:
-                errors[
-                    "Path Errors"
-                ] = "File is missing data_path or contributors_path."
-            else:
-                dir_errors = self._get_ref_errors(rows, "data", schema_version, path)
-                if dir_errors:
-                    """
-                    TODO: there's an issue here (and any other places setting a key
-                    that might be long) where the YAML dumper converts this to a
-                    complex key and adds a ? at the front and inserts a line break
-                    at the next instance of whitespace (in this case, leading
-                    (as x) to be on a following line)
-                    Wrote validation_utils > print_path to try to address this
-                    issue, but needed to deprioritize.
-                    """
-                    errors.update(dir_errors)
+        for path, schema in self.effective_tsv_paths.items():
+            dir_errors = self._get_ref_errors("data", schema, path)
+            if dir_errors:
+                """
+                TODO: there's an issue here (and any other places setting a key
+                that might be long) where the YAML dumper converts this to a
+                complex key and adds a ? at the front and inserts a line break
+                at the next instance of whitespace (in this case, leading
+                (as x) to be on a following line)
+                Wrote validation_utils > print_path to try to address this
+                issue, but needed to deprioritize.
+                """
+                errors.update(dir_errors)
         return errors
 
-    def validation_routine(self) -> dict:
+    def validation_routine(
+        self,
+        report_type: ReportType = ReportType.STR,
+        tsv_paths: Dict[str, SchemaVersion] = {},
+    ) -> dict:
         errors: DefaultDict[str, dict] = defaultdict(dict)
-        for tsv_path, schema_version in self.effective_tsv_paths.items():
-            path_errors = self._validate(tsv_path, schema_version)
-            if path_errors:
-                for key, value in path_errors.items():
+        tsvs_to_evaluate = tsv_paths if tsv_paths else self.effective_tsv_paths
+        for tsv_path, schema_version in tsvs_to_evaluate.items():
+            path_errors = self._validate(tsv_path, schema_version, report_type)
+            if not path_errors:
+                return {}
+            for key, value in path_errors.items():
+                if type(value) is dict:
                     errors[key].update(value)
-        return errors
+                else:
+                    errors.update({key: value})
+        return dict(errors)
 
     def _validate(
         self,
         tsv_path: str,
         schema_version: SchemaVersion,
+        report_type: ReportType = ReportType.STR,
     ) -> Dict[str, Any]:
         errors: Dict[str, Any] = {}
         local_validated = {}
         api_validated = {}
-        rows = self._get_rows_from_tsv(tsv_path)
-        if isinstance(rows, dict):
-            return rows
+        try:
+            rows = read_rows(Path(tsv_path), encoding=self.encoding)
+        except TSVError as e:
+            return {"TSV Errors": e.errors}
         if "metadata_schema_id" not in rows[0].keys():
             logging.info(
                 f"""TSV {tsv_path} does not contain a metadata_schema_id,
                 sending for local validation"""
             )
-            local_errors = self.__get_assay_tsv_errors(
-                schema_version.schema_name, Path(tsv_path)
-            )
+            try:
+                schema = get_table_schema(
+                    schema_version,
+                    self.optional_fields,
+                    self.offline,
+                )
+            except OSError as e:
+                return {
+                    f"{tsv_path} (as {schema_version.table_schema})": {
+                        e.strerror: e.filename
+                    }
+                }
+
+            if schema.get("deprecated") and not self.ignore_deprecation:
+                return {
+                    "Schema version is deprecated": f"{schema_version.table_schema}"
+                }
+
+            local_errors = get_table_errors(tsv_path, schema, report_type)
             if local_errors:
                 local_validated[
-                    f"{tsv_path} (as {schema_version.schema_name}-v{schema_version.version})"
+                    f"{tsv_path} (as {schema_version.table_schema})"
                 ] = local_errors
         else:
             """
@@ -341,7 +371,7 @@ class Upload:
                 return errors
             else:
                 url_errors = self._cedar_url_checks(tsv_path, schema_version)
-                api_errors = self.api_validation(Path(tsv_path))
+                api_errors = self.api_validation(Path(tsv_path), report_type)
                 if url_errors or api_errors:
                     api_validated[f"{tsv_path}"] = url_errors | api_errors
         if local_validated:
@@ -379,7 +409,11 @@ class Upload:
                 errors["Unexpected Plugin Error"] = [e]
         return dict(errors)  # get rid of defaultdict
 
-    def api_validation(self, tsv_path: Path, report_type: ReportType = ReportType.STR):
+    def api_validation(
+        self,
+        tsv_path: Path,
+        report_type: ReportType,
+    ):
         errors = {}
         response = self._cedar_api_call(tsv_path)
         if response.status_code != 200:
@@ -450,8 +484,6 @@ class Upload:
 
     def _check_matching_urls(self, tsv_path: str, constrained_fields: dict):
         rows = self._get_rows_from_tsv(tsv_path)
-        if isinstance(rows, dict):
-            return rows
         fields = rows[0].keys()
         missing_fields = [
             k for k in constrained_fields.keys() if k not in fields
@@ -521,16 +553,18 @@ class Upload:
             return msg if return_str else get_json(msg, error["row"], error["column"])
         return error
 
-    def _get_rows_from_tsv(self, path: Union[str, Path]) -> Union[Dict, List]:
-        errors = {}
+    def _get_rows_from_tsv(self, path: Union[str, Path]) -> List:
+        errors: Dict[str, Any] = {}
         try:
             rows = dict_reader_wrapper(path, self.encoding)
-            return rows
+            if type(rows) is list:
+                return rows
+            errors["TSV Row Errors"] = rows
         except UnicodeDecodeError as e:
             errors["Decode Errors"] = get_context_of_decode_error(e)
         except IsADirectoryError:
             errors["Path Errors"] = f"Expected a TSV, found a directory at {path}."
-        return errors
+        raise ErrorDictException(errors)
 
     def _check_path(
         self,
@@ -553,14 +587,19 @@ class Upload:
                 # TODO: quote field name to match TSV error output;
                 # will break tests
                 errors[f"{metadata_path}, row {i+2}, column {ref}_path"] = ref_errors
+        # TODO: this is all to support other types, should probably be broken out
         else:
-            tsv_ref_errors = get_tsv_errors(
-                schema_name=ref,
-                tsv_path=path,
-                offline=self.offline,
-                encoding=self.encoding,
-                ignore_deprecation=self.ignore_deprecation,
-            )
+            try:
+                schema = get_schema_version(
+                    path,
+                    self.encoding,
+                    self.globus_token,
+                    self.directory_path,
+                )
+            except Exception as e:
+                errors[f"{str(metadata_path)}, row {i+2}, column '{ref}_path'"] = [e]
+                return errors
+            tsv_ref_errors = self.validation_routine(tsv_paths={str(path): schema})
             # TSV located and read, errors found
             if tsv_ref_errors and isinstance(tsv_ref_errors, list):
                 errors[f"{path}"] = tsv_ref_errors
@@ -571,30 +610,19 @@ class Upload:
                 ] = tsv_ref_errors
         return errors
 
-    def __get_assay_tsv_errors(self, assay_type: str, path: Path):
-        return get_tsv_errors(
-            schema_name=assay_type,
-            tsv_path=path,
-            offline=self.offline,
-            encoding=self.encoding,
-            optional_fields=self.optional_fields,
-            ignore_deprecation=self.ignore_deprecation,
-        )
-
     def _get_ref_errors(
         self,
-        rows: list,
         ref: str,
         schema: SchemaVersion,
         metadata_path: Union[str, Path],
     ):
         ref_errors: DefaultDict[str, list] = defaultdict(list)
-        for i, row in enumerate(rows):
+        for i, row in enumerate(schema.raw_rows):
             field = f"{ref}_path"
             if not row.get(field):
                 continue
-            data_path = self.directory_path / row[field]
-            ref_error = self._check_path(i, data_path, ref, schema, metadata_path)
+            ref_path = self.directory_path / row[field]
+            ref_error = self._check_path(i, ref_path, ref, schema, metadata_path)
             if ref_error:
                 ref_errors.update(ref_error)
         return ref_errors