+++ CHANGELOG.md
@@ -36,8 +36,10 @@
 - Update LCMS and add NanoSplits
 - Update descriptions for segmentation masks
 - Add description to codex doc page
-- Started implementing soft assay types for canonical assay names and dir structures
 - Bail earlier in validation if there are errors in metadata/dir/refs
+- Update Antibodies
+- Remove NanoSplits
+- Implemented soft assay types for canonical assay names and dir structures
 
 ## v0.0.15 - 2023-04-04